LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-scroll-position-try-014-expected.txt


FAIL anchor-scroll-position-try-014 assert_equals: Anchored element should be at the bottom of anchor expected 44 but got -56
PASS anchor-scroll-position-try-014 1
FAIL anchor-scroll-position-try-014 2 assert_equals: Anchored element should be at the bottom of anchor expected 59 but got -41
